Consumo del API Rest para los sitios seguros y albergues
