+++ lib/models/zona.dart
@@ -1,0 +1,14 @@
+class Zona {
+  int id;
+  String nombre;
+  String tipo;
+  double cx;
+  double cy;
+  Zona({
+    required this.id,
+    required this.nombre,
+    required this.tipo,
+    required this.cx,
+    required this.cy,
+  });
+}
