+++ pubspec.yaml
@@ -44,6 +44,8 @@ dependencies:
   archive: ^3.3.7
   intl: ^0.18.1
   csv: ^5.0.2
+  http: ^0.13.6
+  connectivity: ^3.0.6
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
